Add oterm-send-command-if-at-prompt.
This change introduces oterm--at-prompt-p which is able to figure out
whether the point is before the prompt, on the prompt or after the
prompt. In the first case it returns nil.

This is useful for keystrokes which should behave differently depending
on where they're sent. For example RET should send the command if the
point is at the prompt or send a normal newline otherwise.