A different approach that avoids send-and-wait.
This change moves away from the practice of using send-and-wait from
commands, especially "send command", which causes problems and delays on
zsh.

The idea is to:
- keep the process mark following the point, so when a command is
called on the command-line, the process mark is at the right place.
- detect prompts using the following heuristic: the pmark after
the string where bracketed paste is enabled is at the end of the
prompt. That seems to work for both bash and zsh, but of course
isn't always available.