Have oterm--modification-hook always delete from the right side.
With this change, deletion always happens from the right side, even when
we're doing a replacement.

This might require up to two moves to figure out the boundaries: one to
figure out how close to BEG we can get and how close to (+ BEG
OLD-LENGTH) we can get. In most cases, we're only deleting or only
inserting, so this code is written so that we would only be one move in
such cases.