Record/replay.
With this change, oterm records what happens during command execution as
text properties then replays that in the post command hook.

This approach avoids having commands being confused by the text in the
buffer not being the same as the text they just inserted. It also avoids
sending unnecessary keystrokes to reproduce intermediate state that's
normally not even noticed. A common example of both is yank followed by
a text property handler which immediately modifies the text just yanked.