Always vertical-motion instead of term-buffer-vertical-motion.
term-buffer-vertical-motion has problems - but then it's hardly ever
used in normal operation since terminals are usually displayed in a
window.

This change uses oterm's window to run vertical-motion when possible.
This is (surprisingly) supported by vertical-motion and a reasonable
thing to do since the terminal width and height come from oterm's
window.

There are two cases when window dimension don't necessarily match
term-width and term-height, which could cause issues:

- what if there are multiple windows displaying oterm with
different dimensions?

- this change falls back to using the currently selected window to run
vertical motion on if oterm is not displayed in any windows; the
dimensions are unpredictable.