Actualize possible prompts before a positional command.
A positional command is defined as a command that modifies the
command-line string in a way that depends on the position of the cursor.

With this change, when there are no recognized prompt at the current
line and if the point is not on pmark before a positional change and
it's on a possible prompt, misTTY assumes that the user meant to move
the cursor to that point, and decides that the current line is actually
a prompt, then move the cursor.

If movement isn't actually supported, the user will see [OC and [OD as
if they tried to move the cursor themselves.

There are now two kinds of prompts:
 - prompts detected based on backeted-paste, such prompts can be
 multiline
 - possible prompts actualized when the user sent a specific command,
 such prompts cannot be multiline