Make sure to sync the work buffers after syncing was inhibited
Before this change, if something changed in the term buffer while
running a command, the work buffer would only be refreshed the next time
something is sent from the process - assuming that sync is not disabled
again when that happens.

This change remembers when term-to-work is inhibited and make sure the
post-command hook calls term-to-work at the end when necessary.