Disable process output only while collecting modifications.
Before this change, mistty would allow updating the term buffer but
inhibit the copying of changes to the term buffer to the work buffer:
 - during a command
 - while replaying

This caused issues because there's more to just copying to the term
buffer that happens during the process filter.

This also caused the screen to be unrefreshed during commands.

Not accepting process output at all allows all the normal processing to
happen, just later, at an appropriate time.

Only disabling while collecting changes, just after the first change was
made, gives us the stability we need just when we need it.