Go back to point after applying the last modification.
If the point was inside the modification area (as is usually the case),
move pmark to the goal position and have the point be set from that.
This allows taking into account modifications made by the process or
term.el (newlines).

This means that moving the pmark to the point after a command shouldn't
happen when modifications are applied, as it would interfere.