Update prompt-detection heuristic based on bracketed paste.
Before this change, a prompt was set at the pmark in the string that
turned on bracketed paste. This generally worked OK, but in some cases
the filter was called twice: once to turn on bracketed paste and once to
write the prompt. This happened with fish.

This change doesn't rely so much on grouping by remembering when
bracketed paste was turned on and using the first acceptable pmark after
that as prompt.