Inhibit term-to-work while handling modifications.
This change switches back to disabling term-to-work while handling
modification until the very end, when the string containing the last
modification is sent. This way, we're much less likely to see
intermediate state with the modification turned off.

This allows multiple calls to hippie-expand to work, as if it sees an
intermediate state with its modifications turned off, it'll lose its
marker and won't be able to delete the output of its previous command.