Fix text-read-only error when leaving fullscreen mode.
Before this change, term-to-work would fail with a read-only error
triggered on the term buffer just before leaving fullscreen mode.

This change turns on inhibit-read-only for the whole of term-to-work and
moves away from detecting text-read-only errors from
replace-buffer-contents in favor of saving the prompt and detecting
changes there manually.