Custom term-command-function that works under fish.
Fish adds ESC [ K before the newlines of strings output by fish_prompt,
which breaks the term-command-function, as it uses a newline as end
separator.

This new implementation gets a string length first, so whatever garbage
is added between the end of the string and the newline is going to be
ignored. It also takes a hostame - just for verification, for now,
though it would be possible to build a tramp filename based on it.