Deal with empty prompts, fix issue with zsh and mistty-prompt-id.
This change updates the prev-next prompt commands to rely on
mistty-prompt-id, which allows distinguishing prompts that are right
next to each other.

It also fixes an issue with zsh and the properties-based prompt
detection: as zsh only turns bracketed-paste mode after writing the
prompt, we couldn't see anything with a prompt-id property until the
user typed something in the prompt - which defeats the purpose. This
change switches to setting this property on everything on the current
line before the point where bracketed-paste was turned on, which assumes
a non-empty prompt on the current line for zsh.

This change highlights an issue with restrictions: accessing a text
property outside of the restricted region strictly fails, which causes
breakages when there is a restriction.