Don't destroy the prompt field when the prompt changes.
Before this change, the mistty-cmd-start-marker was reset to
mistty-sync-marker whenever anything was modified within the prompt.
This mean that the prompt was lost under bash when using C-r or under
fish when the prompt is refreshed asynchronously.

With this change, the code now relies on replace-buffer-contents to move
mistty-cmd-start-marker appropriately when the prompt is modified. It
seems to be doing a rather good job; why interfere with it?

Similarily, this change reverts the saving of undo-list. This was never
tested and might have been doing more harm than good.