Introduce C-q <any> that send everything to the terminal.
Before this change, some keys could be sent to the terminal using C-c
<key>, but it was never complete.

This change introduces C-q so that all keys that are known can be sent
to the terminal. This is only active while on the synced region, usually
the prompt.

This change leaves some important C-c <key> shortcut available in the
whole buffer, such as C-c C-c to send a C-c.