Don't trust prompt detection.
There are cases where the heuristic MisTTY uses for prompt detection
fails. Luckily, the exact position of prompts is just a nice to have
except that:
 (a) prompts are marked read-only so if the detection is
 over-enthusiastic, the result may be an uneditable command-line
 (b) beginning-of-line relies on prompt detection to work, or it goes
 to the real beginning of the line
 (c) next/previous prompt rely n prompt detection to work, or they go
 to the beginning of the interactive region

To deal with (a), this change stops setting the read-only property. Now
you can edit prompts; though the edits will normally just be reverted
right away.

To deal with (b), this change switches to sending C-a/C-e and letting
the shell decide where the beginning or end of line is; the results are
generally much better.

This change does nothing to address (c).