Making change replay asynchronous using generators.
This change transforms mistty--replay-modifications into a generator,
which generates strings to send to the process and is called again once
the process has received input, presumably the response.

In addition, this change switches to using idle timer both for
post-process-command-1, but also for processing "responses" from the
process. In the first case, this de-prioritize post-process-command-1
and avoid freezing Emacs, in the first case, this delays processing the
response enough to give time for multiple responses sent one immediately
after the other to be buffered, just like waiting for response did
before.

This change allows executing commands while changes are being processed
by the process, but it doesn't yet attempt to make sure only safe
changes are made.