Add a timeout, in case the shell doesn't react to a sequence.
Before this change, it would be easy to get into a situation where the
replay generator would hang, because the shell would return no answers
whatsoever, such as for a movement left when already at the leftmost
position.

This change introduces a timer for timing out if there's no answer after
some time.