Define a fringe bitmap and merge mistty-fringe-face and -margin-face.
Before this change, the fringe bitmap wasn't really used, instead the
background and foreground color were set to fill the whole fringe,
whatever its width was.

With this change, the purple bar in the fringe is defined to be 3 pixel
width and take the whole height. This means that the hack of defining
both background and foreground color can go and we can use the same face
for both fringe and margin (in the terminal).