C-e goes immediately to the cursor.
When the prompt is empty, sending C-e has no effect, as doesn't move the
prompt. This makes C-e useless to go back to the prompt, if the pointer
is on the prompt.

With this change, typing C-e goes to the cursor, and if the cursor never
moves, something useful has happened at least.