Fix strange term behavior when leaving fullscreen mode.
Before this change, MisTTY would set mistty-fullscreen-map as local map
when entering fullscreen mode. This caused term.el to think it was it
line mode, because the map is what is used to detect the mode:

  (defmacro term-in-char-mode () '(eq (current-local-map) term-raw-map))
  (defmacro term-in-line-mode () '(not (term-in-char-mode)))

This change sets the local buffer value of term-raw-map in term-mode
buffers created by MiTTY to mistty-fullscreen-map. This way, the right
map is set and term correctly understands that this is the char-mode
map.