Set process window size to term-width and term-height.
Before this change, the process window size was set automatically when
the term-mode buffer was created, which meant that it could be different
from term-width and term-height - especially when using margin.

This might have been the cause of the width as understood by fish
sometimes being 1 char too large.

With this change, the process window size is set early on, just after
starting the process.