Add support for fish multiline prompts.
Before this change, MisTTY's was unable to move the cursor to follow the
point in fish multiline prompt, because the approach taken was
completely confused by indent and whitespaces after the command.

This change introduce specific support for fish multiline prompts when
moving between lines, to detect them and then use the up/down arrows to
move between lines. Up/down is normally avoided as up/down normally go
up/down in history, but in fish multiline prompts, they behave normally.

Behavior when it comes to going to spaces after the command words is
still spotty, as these spaces might or might not reachable; there's no
way to tell.