Fix fish multiline handling of trailing ws and empty lines.
Before this change, it would not be possible to correctly move to or
from a trailing whitespace; they just didn't exist for the algorithm. To
avoid that, this change removes whitespaces after having computed the to
and from column.

Before this change, going through an empty line we'd go through a column
with a negative number; that didn't work. If a line is empty, assume
column 0 is usable.