Interrupt replay when reseting or going fullscreen.
Before this change, replay were not cleanly interrupted: they would
continue after a reset, despite the forced refresh, and they would be
reset when going to fullscreen, but the changeset would remain.

This change adds an unwind form to replay generator that clears the
changeset and clears inhibit-refresh. It also makes sure that iterators
are closed when the queue is cancelled. This allows for clean
cancellation in both cases.