Fix test-mistty-term-buffer-exits.
Before this change, there was a race condition in
test-mistty-term-buffer-exits: if the term buffer died before entering
mistty-wait-for-term-buffer-and-proc-to-die, mistty-proc could already
have been set to nil.

This change stores the value of mistty-term-buffer and mistty-proc at a
point where we know they already exist, so they won't be nil when we
need them.