Stop setting the 'field text property on detected prompts.
Setting the 'field text property is useful to indicate to
beginning-of-line and other movement commands that there's a boundary.
However, beginning-of-line is sent to the terminal as C-e, and
boundaries are enforced by the shell itself.
So, in this situation, the text property is not that useful.

In addition, setting the 'field has bad consequences: when prompt
detection is over-enthusiastic, movement commands behave unexpectedly.

Better leave it off entirely and just continue with the strategy of not
relying on prompt detection for anything important.

This change rewrite the next/previous command to rely on the text
property 'mistty = 'prompt instead of 'field = 'mistty-prompt.