Stop relying on detected prompt to detect fish multiline.
Before this change, fish multiline mode, which uses up/down arrows,
would not trigger if the detected prompt was not just the right size.

With this change, fish multiline mode doesn't rely on the prompt
anymore, but rather on a common and non-empty indent. This works even if
the detected prompt is too large or empty. The risk is that it might
work when not editing fish multiline. Luckily, outside of fish,
multiline either rely on fake lines added by the terminal or rely on a
secondary prompt.