Don't bother inserting content that's already there.
When replaying modifications, we often have cases where parts of the
inserted string is the same as the string that was just deleted. This
typically happens when using hippie-complete.

This change adds some code to detect when the beginning of the inserted
string is already there and just skip inserting that part.