Mark and skip fake spaces added by the terminal when moving around.
(move-to-column) adds spaces when necessary, this means that spaces
appear on the terminal that correspond to nothing but a path for the
cursor to go through.

Such spaces confuses the algorithm for moving the cursor to a desired
position: as these spaces don't really exist from the program's point of
view, they can't be reached with the cursor.

This change addresses the problem by marking such spaces with a special
text property that's later detected by mistty--distance-on-term, so they
aren't counted when generating movements.