Remove cursor argument from mistty--move-generator.
Having a cursor argument makes no sense; it always start at the cursor.
Also, storing the cursor position when the generator is created will
cause issues if the generator is queued, as the cursor might move
between the time the generator is created and the time it is executed.

Also rename it move-cursor-generator