Use fire-and-forget for self-inserting characters.
In general, MisTTY waits for every set of characters sent to the
terminal to be echoed back to it before sending the next one. This has
the advantage of forcing actions to be strictly sequential and avoid
race conditions. This can be slow, however, especially on remote
connections.

This change attempts to improve the situation slightly, without taking
much risks, by dropping the wait when sending out self-inserting
characters with mistty-send-key. This lets them be buffered by the
program, when necessary.