Move the cursor back to point right away after replay.
Before this change, the cursor was moved back to its initial position
only after having received the result from the last insertion/deletion.
This is unnecessary, since we know where the cursor is going to be at
the end of the insertion/deletion.

This change puts the string that moves the point into the same string as
the one that includes the last insertion/deletion.

This change also recovers the point if it was between mistty-sync-start
and intervals-start.