Fix cursor handling after replay.
Before this change, after the last change was replayed,
mistty--replay-generator attempted to move the cursor to the initial
position.

That didn't always work too well, especially when the cursor was after
the change. Even if the change went perfectly as planned, invisible
spaces could have added, making the remembered position incorrect.

This change gives up on trying to figure out where the point should be
after a change and moving the cursor to it. It relies instead on
replace-buffer-contents to do the right thing when it comes to point and
then attempt to move the cursor to the point. This has the additional
advantage of conserving the position of points that is just not
reachable with cursors.