Create a temp buffer (backstage) for replay to work on.
Before this change replay had stability issues due to unexpected changes
to the term buffer, such as the addition of trailing whitespaces,
invisible but enough to invalidate positions.

This change relies on markers to avoid this problem. As markers cannot
work on the term buffer and the work buffer must remain frozen until
everything has been replayed, this change introduces a temporary
"backstage" buffer which is kept up-to-date during replay using
replace-buffer-content, just like the work buffer. Replay works fully on
this backstage buffer; this is easier than the previous approach which
worked on the term buffer, but kept using work buffer positions.