Simplify the test for moving point after terminal buffer dies.
Before this change, the point was compared with the cursor position,
which requires access to the work buffer, the term buffer and the
process. This is not reliable while the system is being brought down.

This change simplify the condition for moving the cursor to the end. The
result is more aggressive, but should boil down to: if the cursor is at
the prompt, move it, otherwise leave it alone.