test-mistty-term-buffer-exits: check point after waiting.
The flakiness of the test might be due to a race condition: "finished"
appears before the point is moved, so we're not sure the point is ready.
The point is moved before the term buffer is allowed to die, so moving
the check after mistty-wait-for-term-buffer-and-proc-to-die should make
the test more reliable.