Avoid moving to right prompts or yanking their value.
With this change, the empty spaces added for creating right prompts as
well as the right prompt itself are not yanked if copied or killed.

Also, the right prompt is not accessible if mistty-skip-empty-spaces is
enabled.

This works only as long as there is a "gab" between the command-line and
the right prompt.

Tested with fish right prompts.