Fix for cursor moving to the next line after a new prompt.
Before this change, with mistty-skip-empty-spaces set, fish and a right
prompt, sometimes the cursor would end up after the prompt line at a new
prompt, because in some intermediate state the cursor ended up inside of
a mistty-skip.

This change always respects a position that was set by following the
cursor, usually after a refresh, as opposed to a position that was set
by moving the point.