Run tests in strict mode.
With this change, while in tests, any time a sequence is sent to the
terminal by a generator with a function to test its effect, the effect
MUST eventually be found in the terminal.

This isn't something that can be turned on generally, as there are cases
where the desired effect cannot even by replicated by the terminal, such
as if a text is inserted or deleted outside the point.

This is useful in tests, however, as it makes sure that the conditions
work and that there are no surprise timeouts.

This change also extends the conditions used to detect inserted or
deleted text in the terminal to keep them working even if the inserted
text is slightly modified.