Don't skip completely empty lines.
Before this change, completely empty lines in a multiline prompt were
inaccessible with mistty-skip-empty-spaces and were in general not
counted by mistty--distance.

With this change, such empty lines count for 1 in mistty-distance and
are not completely skipped over with mistty-skip-empty-spaces.

There needs to be some place where a cursor goes *somewhere* in an empty
line on a multiline prompt, for the line to be editable, but nothing
indicates that. This change always puts the cursor at the end of the
line, which is not always correct.