Fix distance computation with partially skipped region.
With this change, a skipped region counts for the distance only if it is
complete, going from beginning to end. It doesn't count if started after
the beginning or if the end is not reached.

This makes more sense for distances between points that don't start or
end in a reachable position; everything that's not reachable doesn't
count.