On empty lines, put the cursor at the prev line indentation.
Before this change, the cursor was put, arbitrarily, at the end of an
empty line.

With this change, the cursor is put at the position that correspond to
the previous line indentation, if possible. This looks better, because
it reproduces what fish does.

The position is still arbitrary and the cursor likely to move as
something gets typed into the line.