Fix distances that start or end on an unreachable point.
Before this change, distance computation was off when the start or end
of the distance was unreachable. What should have happened was that
crossing the skipped region to go from unreachable point to go to the
closest reachable point (or the reverse) would count for 0, but it
counter for 1.

This change fixes this issue and adds a test specifically for that.