Remove the special value 'cursor from mistty--last-pos.
Before this change, mistty--last-pos was set with a special 'cursor
value when the point is moved to the cursor, which mistty--cursor-skip
was supposed to respect, but it never really did.

This change replaces that with a clearing of last-pos, so that movement
duo to going to the cursor don't look like the result of normal move
commands.