Turn off editing when shells enter special modes.
Before this change, when fish was in backward-search mode, the result
list would go crazy as MisTTY would try to move the cursor, but
couldn't.

With this change, a set of regexps detect backward-search mode in
different shells and turn off:
 - replaying of changes made to the buffer
 - cursor tracking
So behavior stays sane.

More work is needed to build a comprehensive set of regexps for this
mode for different shells.