Stop sending unnecessary left/right or right/left pairs.
Before this change, mistty--move-horizontally-str would add an no-op
left/right or right/left move at the end to force the shell to send
something in case the cursor was stuck. This was meant to avoid timing
out when the shell just sends nothing because the cursor position is not
available.

This caused issues:
 - if one of the left/right or right/left pair worked, the final
 position was sometimes wrong
 - mistty--yield attempting to detect whether the goal position fired
 too early, because we still had things to process. This had
 unpredictable effects in tests.

With this change, mistty--move-horizontally-str never sends any
unnecessary moves, and we rely on timeouts to deal with that reasonably.
This is slower, but more predictable.