Track old sync mark positions that are still on the screen.
This change introduces mistty--sync-history, a list of marker
pairs (work buffer/term buffer) that correspond to previous positions of
the sync mark that are still on the screen and whose content hasn't been
modified by the user.

The goal of this is to improve recovery when the sync mark was moved too
quickly in future changes.