Reset up to an older sync mark, when possible.
Before this change, when a modification on the terminal was detected
above the sync mark, the sync mark was reset to the beginning of the
terminal screen, on the term buffer, and to the current position on the
work buffer.

In the case where the terminal content was just slightly modified, this
resulted in a great deal of duplication on the work buffer and was
generally very confusing.

With this change, when a modification on the terminal is detected,
MisTTY tries to reset the sync mark to an older point above that
modification. This doesn't work when the whole terminal is cleared, but
does work if something is modified just above a sync mark - usually a
sync mark that was set after detecting and realizing a prompt that
turned out not to be real.