Inhibit undo when refreshing and replaying changes.
Before this change, some parts of refreshing and replay changes were
included into the buffer undo list: the text properties. This had
surprising effects.

With this change, changes made by the remote process is left out of the
undo list. This includes, for now, changes resulting from the user
sending a key to the process.

Only change made Emacs-side and later replayed are currently undoable.