Avoid kill-line being confused by fake newlines.
Before this change, when calling kill-line on a line longer than the
window width, kill-line would only apply on the portion of the line up
to the window edge. This is because term.el adds fake newlines there.

This change tweaks the text properties of these fake newlines to allow
kill-line to go through them and delete the whole real line. It make
these newlines invisible - but then they weren't really visible to begin
with since they're always at a window edge.