Make self inserted keys and delete keys undoable.
Before this change, undo would only work on replayed changes, changes
that were applied on the buffer first, then sent to the terminal. It
would not apply to keys that are sent directly to the terminal program.

This change manually adds self inserted keys DEL and C-d to the undo
list, so they can be undone as they would be in a normal buffer.