Undo supports deleting text that was just typed.
Before this change, undo didn't generate the right undo entry if text
that was just typed was deleted before the text that was typed was
displayed on the terminal, when typing fast.

This change stores the text that was just typed for the duration of the
undoable key chain, so it knows whether DEL applies to just typed text
or text that's in the buffer.