Consider cursor-to-point successful once at (point).
Before this change mistty--cursor-to-point-generator would store the
position of (point) at the beginning of the call and expect to reach
that position at the end of the call. That didn't take into account any
changes that would be applied to the buffer before the point.

With this change, mistty--cursor-to-point-generator call (point) again
when comparing the cursor against the goal.