Remove left/right hack after bracketed pasting text.
When pasting text using bracketed paste, the resulting region is often
left highlighted by the shell, which looks bad.

Before this change, MisTTY would add a left/right movement just after
pasting the text, which would de-activate the region. This caused issues
with the logic that detects when a change is successful, however, as the
change looks "finished", but is then re-drawn because of the cursor
movement. The result was unpredictable, especially in test.

This change removes the left/right movement entirely, which might leave
some text highlighted which shouldn't. This might have visible effect -
though it's likely limited as cursor often move after a paste anyway.

This change is a way of judging how bad this highlighting look in
practice and decide whether it's worth introducing other workarounds.
One workaround would be to split the text into pasted and normal text
and make sure that the last pasted text is not bracketed.