Have mistty-test-fullscreen-live-buffer-p clean itself up.
mistty-test-fullscreen-live-buffer-p tended to leave the term-mode
buffer around. This is because, since the work buffer was killed while
in fullscreen mode, it would not cause the term buffer to be killed as
well, since the term buffer is the main buffer (the one the user sees.)

This change switches back to normal mode before the end of the test.