Update the way replayed insertion and deletion are detected.
Before this change replayed insertion and deletion were detected with a
regexp starting at the beginning of the change that ignored spaces.

This had unpredictable effects when inserted or deleted test contained
spaces. The beginning position was also sometimes not moved exactly as
it should have by replace-buffer-content, so being strict about it was
not a good idea.

This change switches to an approach that ignores fake newlines and
term-skip spaces, but leaves other spaces and newline as significant, by
removing the before building or applying the regexp.

To avoid issues with imprecision in the beginning position, this change
compare everything between the beginning and end of
the (non-term-line-wrap'ed) line when detecting changes.

Further, to avoid starting moving the cursor after the text has been
written but before the cursor has moved to its stable, eventually
user-visible position, this change new checks that the cursor is at the
expected position, at the end of the inserted or deleted region.

This should avoid timeouts and make the tests less flakey.