In tests, check that cursor is at the end of the inserted text.
Before this change, tests would sometime start making changes with the
cursor at a temporary position, while it's in the process of writing
inserted text to the terminal. This made tests break.

To avoid that, this change has the tests waiting for cursor to be at the
end of the text that was just inserted, just like when replaying.

This should make tests less flaky in general.