Update mistty-test-sync-history-autoreset to look at the history.
Before this change, mistty-test-sync-history-autoreset checked prompt
detection to figure out whether there were older sync markers in the
text. This was overly indirect and caused some flakiness when the prompt
wasn't detected properly.

This change check the content of the history after the reset instead,
which is what was meant to be tested to begin with.