When moving the point, ignore spaces with mistty-skip t.
MisTTY can't figure out where the cursor will be with any certainty in
multi-line prompts with empty lines, as the position might depend on the
command - so figuring it out would require understanding the language.

This isn't much of a problem in practice, as the position is set by the
shell once the user types anything.

It's only a problem when moving the cursor, as the cursor might not want
to move exactly where MisTTY expects.

This change works around that by accepting any position on an empty
line, or, put another way, accept as the same position two positions
separated only by spaces with mistty-skip t.