Have the "Fullscreen mode ON" message lookup the key binding.
Before this change, when entering the fullscreen mode, the default key
binding for switching between terminal and scrollback buffers was
displayed.

With this change, the actual key binding is looked up in the keymaps and
displayed.

This change also simplifies toggling between buffers by having just one
command instead of two.