Send C-a/C-e only when on a prompt.
Before this change, mistty-send-beginning-of-line and
mistty-send-end-of-line would send C-a/C-e whenever the pointer was in
the synced region.

This meant that if the prompt wasn't moved quickly enough, such as
happens with possible prompts (python), C-a/C-e would move the pointer
to another line entirely.

This change tries to be a bit more clever and chose dynamically whether
to call beginning-of-line/end-of-line or send C-a/C-e.