Fix end prompts with possible prompts.
When mistty-send-command is called, it set things up so that MisTTY
puts a new sync mark after the next newline.

Before this change, this worked with detected prompts (prompts with
bracketed paste on) but not with possible prompts, because this wasn't
tested. This change fixes that and make it work as it should.