Have mistty--line-indent ignore mistty--cmd-start-marker.
The indent of the prompt line is now 0. This should only have an effect
in multi-line prompt, when figuring out where the cursor should go on
that empty line.

Before this change, the cursor would sometimes go below
the prompt, sometimes go to the left column. With this change, it always
goes to the left column.