Fix flakiness in mistty-test-set-ps1.
Before this change, there was a race condition in mistty-send-text
output detection when sending out the PS1, because it happens so early,
while the shell is still initializing.

This change turns off output detection for this case; it's enough to
rely on detecting the prompt that is defined by setting PS1.