Avoid creating possible prompts during process output.
Before this change, MisTTY would sometimes get confused during process
output and create prompts where there wasn't one if the user used the
beginning of line, end of line or return command on that output, even
though the point was nowhere near the cursor.

This change adds additional safeguards to avoid that. Now both the point
and cursor need to be in the possible prompt for it to be realized.