Include mistty.texi instead of mistty.info.
mistty.info is very much unstable, that is, even if the documentation is
not modified, the generated mistty.info is always different:
 - "Generated by Sphinx <version>" changes when using a different
 version
 - it contains the generation date
 - it contains auto-generated refs that always change

The last problem is the worse. To solve that, this change switch to
sumbitting mistty.texi instead.

The other two problems are addressed in this change by post-processing
the file in Eldev.