Fix initial and modified terminal size.
Before this change, the initial terminal width and height was taken from
the currently selected window, which might or might not be the window
that ends up displaying the buffer.

This change fixes the problem by making sure the buffer is displayed in
a window before the process is started, in mistty-create.

Before this change, in mistty--window-size-change, the margin was
substracted from the height instead of the width.

This change fixes that as well and introduces tests to cover
mistty-create, initial size and modified size.