Integrate mistty--interact-wrap-accept into mistty--interact-return.
Before this change, pretty much all calls to mistty--interact-return
were preceded with (setq accept-f (mistty--interact-wrap-accept ...))
and the next callback started with a call to accept-f.

This change streamlines this by integrating it into
mistty--interact-return, which takes care of wrapping the lambda and
calling it, so the callback can just continue the work without worrying
about accept-f.