Fix and test mistty--swap-buffer-in-windows.
Before this change, mistty--swap-buffer-in-windows attempted to modify
the prev-buffers list of windows, but was doing it wrong, leaving
markers that didn't match the buffer.

With this change, mistty--swap-buffer-in-windows makes sure to keep the
prev-buffer lists as they are, even when it swaps the buffers.

This change also includes a test for mistty--swap-buffer-in-windows.