Fix deletion replay with non-bracketed-paste prompt.
When replaying deletion on a prompt that doesn't use bracketed-paste,
the end-of-line check of mistty-on-prompt-p might get things wrong, as
the cursor might be on a position that's after the current line on the
work buffer.

This change covers the case by checking for the position in both
buffers to decide whether it is on the prompt.