Enqueue changes made before sending a string to the terminal.
Before this change, sending a string to the terminal was added to
mistty--queue and executed in order, but prompt detection, going to the
cursor and setting mistty-goto-cursor-next-time was executed right away.

When the queue was busy, that meant that these actions would be executed
out-of-order. This was especially visibly with
mistty-goto-cursor-next-time, which would be set during a replay and
prevent the goto-point interaction from doing anything.

This change puts everything related to sending a string to the terminal
into the queue, creating a custom mistty--interact for each of them.