Detect completion-in-region as a long-running-command.
Before this change, the replay and refresh mechanism could confuse or
abort completions, depending on the UI.

To avoid that, this change detects when completion-in-region is active,
and treats that as a long-running-command, like filling in a template.
This means that replay and refresh is turned off for the duration of the
completion. This works for completion UIs that turn on
completion-in-region mode.

Notably, this allows corfu UI to work when called explicitly, for
issue #5

This change does nothing for auto-completion.